+++ cmd/gobench-cable/main.go
@@ -59,7 +59,12 @@ func main() {
 	controller := gobench.NewController(metrics)
 
 	appNode := node.NewNode(controller, metrics, &config.App)
-	appNode.Start()
+	err = appNode.Start()
+
+	if err != nil {
+		log.Errorf("!!! Failed to initialize application !!!\n%v", err)
+		os.Exit(1)
+	}
 
 	var disconnector node.Disconnector