+++ node/config.go
@@ -6,9 +6,11 @@ type Config struct {
 	PingInterval int
 	// How ofter to refresh node stats (seconds)
 	StatsRefreshInterval int
+	// The max size of the Go routines pool for hub
+	HubGopoolSize int
 }
 
 // NewConfig builds a new config
 func NewConfig() Config {
-	return Config{PingInterval: 3, StatsRefreshInterval: 5}
+	return Config{PingInterval: 3, StatsRefreshInterval: 5, HubGopoolSize: 16}
 }