+++ node/hub_test.go
@@ -1,6 +1,9 @@
 package node
 
 import (
+	"fmt"
+	"math/rand"
+	"sync"
 	"testing"
 	"time"
 
@@ -8,13 +11,15 @@ import (
 )
 
 func TestUnsubscribeRaceConditions(t *testing.T) {
-	hub := NewHub()
+	hub := NewHub(2)
+	node := NewMockNode()
 
 	go hub.Run()
 	defer hub.Shutdown()
 
-	session := NewMockSession("123")
-	session2 := NewMockSession("321")
+	session := NewMockSession("123", &node)
+	session2 := NewMockSession("321", &node)
+	session3 := NewMockSession("213", &node)
 
 	hub.addSession(session)
 	hub.subscribeSession("123", "test", "test_channel")
@@ -22,24 +27,21 @@ func TestUnsubscribeRaceConditions(t *testing.T) {
 	hub.addSession(session2)
 	hub.subscribeSession("321", "test", "test_channel")
 
+	hub.addSession(session3)
+	hub.subscribeSession("213", "test", "test_channel")
+
 	hub.Broadcast("test", "hello")
 
-	done := make(chan bool)
-	timer := time.After(100 * time.Millisecond)
+	_, err := session.ws.Read()
+	assert.Nil(t, err)
 
-	go func() {
-		<-session.send
-		<-session2.send
-		done <- true
-	}()
+	_, err = session2.ws.Read()
+	assert.Nil(t, err)
 
-	select {
-	case <-timer:
-		t.Fatalf("Session hasn't received any messages")
-	case <-done:
-	}
+	_, err = session3.ws.Read()
+	assert.Nil(t, err)
 
-	assert.Equal(t, 2, hub.Size(), "Connections size must be equal 2")
+	assert.Equal(t, 3, hub.Size(), "Connections size must be equal 2")
 
 	go func() {
 		hub.Broadcast("test", "pong")
@@ -49,33 +51,26 @@ func TestUnsubscribeRaceConditions(t *testing.T) {
 
 	go func() {
 		hub.Broadcast("test", "bye-bye")
+		hub.RemoveSession(session3)
+		hub.Broadcast("test", "meow-meow")
 	}()
 
-	timer2 := time.After(2500 * time.Millisecond)
-	go func() {
-		<-session2.send
-		<-session2.send
-		<-session2.send
-		<-session.send
-		done <- true
-	}()
-
-	select {
-	case <-timer2:
-		t.Fatalf("Session hasn't received any messages")
-	case <-done:
+	for i := 1; i < 5; i++ {
+		_, err = session2.ws.Read()
+		assert.Nil(t, err)
 	}
 
 	assert.Equal(t, 1, hub.Size(), "Connections size must be equal 1")
 }
 
 func TestUnsubscribeSession(t *testing.T) {
-	hub := NewHub()
+	hub := NewHub(2)
+	node := NewMockNode()
 
 	go hub.Run()
 	defer hub.Shutdown()
 
-	session := NewMockSession("123")
+	session := NewMockSession("123", &node)
 	hub.addSession(session)
 
 	hub.subscribeSession("123", "test", "test_channel")
@@ -83,54 +78,39 @@ func TestUnsubscribeSession(t *testing.T) {
 
 	hub.Broadcast("test", "\"hello\"")
 
-	timer := time.After(100 * time.Millisecond)
-	select {
-	case <-timer:
-		t.Fatalf("Session hasn't received any messages")
-	case msg := <-session.send:
-		assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"hello\"}", string(msg.payload))
-	}
+	msg, err := session.ws.Read()
+	assert.Nil(t, err)
+	assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"hello\"}", string(msg))
 
 	hub.unsubscribeSession("123", "test", "test_channel")
 
 	hub.Broadcast("test", "\"goodbye\"")
 
-	timer = time.After(100 * time.Millisecond)
-	select {
-	case <-timer:
-	case msg := <-session.send:
-		t.Fatalf("Session shouldn't have received any messages but received: %v", string(msg.payload))
-	}
+	_, err = session.ws.Read()
+	assert.NotNil(t, err)
 
 	hub.Broadcast("test2", "\"bye\"")
 
-	timer = time.After(100 * time.Millisecond)
-	select {
-	case <-timer:
-		t.Fatalf("Session hasn't received any messages")
-	case msg := <-session.send:
-		assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"bye\"}", string(msg.payload))
-	}
+	msg, err = session.ws.Read()
+	assert.Nil(t, err)
+	assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"bye\"}", string(msg))
 
 	hub.unsubscribeSessionFromAllChannels("123")
 
 	hub.Broadcast("test2", "\"goodbye\"")
 
-	timer = time.After(100 * time.Millisecond)
-	select {
-	case <-timer:
-	case msg := <-session.send:
-		t.Fatalf("Session shouldn't have received any messages but received: %v", string(msg.payload))
-	}
+	_, err = session.ws.Read()
+	assert.NotNil(t, err)
 }
 
 func TestSubscribeSession(t *testing.T) {
-	hub := NewHub()
+	hub := NewHub(2)
+	node := NewMockNode()
 
 	go hub.Run()
 	defer hub.Shutdown()
 
-	session := NewMockSession("123")
+	session := NewMockSession("123", &node)
 	hub.addSession(session)
 
 	t.Run("Subscribe to a single channel", func(t *testing.T) {
@@ -138,13 +118,9 @@ func TestSubscribeSession(t *testing.T) {
 
 		hub.Broadcast("test", "\"hello\"")
 
-		timer := time.After(100 * time.Millisecond)
-		select {
-		case <-timer:
-			t.Fatalf("Session hasn't received any messages")
-		case msg := <-session.send:
-			assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"hello\"}", string(msg.payload))
-		}
+		msg, err := session.ws.Read()
+		assert.Nil(t, err)
+		assert.Equal(t, "{\"identifier\":\"test_channel\",\"message\":\"hello\"}", string(msg))
 	})
 
 	t.Run("Successful to the same stream from multiple channels", func(t *testing.T) {
@@ -153,23 +129,18 @@ func TestSubscribeSession(t *testing.T) {
 
 		hub.Broadcast("test", "\"hello twice\"")
 
-		done := make(chan bool)
 		received := []string{}
 
-		go func() {
-			received = append(received, string((<-session.send).payload))
-			received = append(received, string((<-session.send).payload))
-			done <- true
-		}()
+		msg, err := session.ws.Read()
+		assert.Nil(t, err)
+		received = append(received, string(msg))
 
-		timer := time.After(100 * time.Millisecond)
-		select {
-		case <-timer:
-			t.Fatalf("Session hasn't received enough messages. Received: %v", received)
-		case <-done:
-			assert.Contains(t, received, "{\"identifier\":\"test_channel\",\"message\":\"hello twice\"}")
-			assert.Contains(t, received, "{\"identifier\":\"test_channel2\",\"message\":\"hello twice\"}")
-		}
+		msg, err = session.ws.Read()
+		assert.Nil(t, err)
+		received = append(received, string(msg))
+
+		assert.Contains(t, received, "{\"identifier\":\"test_channel\",\"message\":\"hello twice\"}")
+		assert.Contains(t, received, "{\"identifier\":\"test_channel2\",\"message\":\"hello twice\"}")
 	})
 }
 
@@ -183,4 +154,95 @@ func TestBuildMessageString(t *testing.T) {
 	expected := []byte("{\"identifier\":\"chat\",\"message\":\"plain string\"}")
 	actual := buildMessage("\"plain string\"", "chat")
 	assert.Equal(t, expected, actual)
+}
+
+type benchmarkConfig struct {
+	hubPoolSize       int
+	totalStreams      int
+	totalSessions     int
+	streamsPerSession int
+	payload           string
+}
+
+func BenchmarkBroadcast(b *testing.B) {
+	configs := []benchmarkConfig{}
+
+	poolSizes := []int{128, 16, 2, 1}
+	streamNums := [][]int{
+		{1000, 10},
+		{100, 10},
+		{10, 3},
+	}
+	sessionsNum := 10000
+	payload := "\"A quick brow fox bla-bla-bla\""
+
+	for _, streamNum := range streamNums {
+		for _, poolSize := range poolSizes {
+			configs = append(configs, benchmarkConfig{poolSize, streamNum[0], sessionsNum, streamNum[1], payload})
+		}
+	}
+
+	for _, config := range configs {
+		b.Run(fmt.Sprintf("%v", config), func(b *testing.B) {
+			broadcastsPerStream := b.N / config.totalStreams
+			messagesPerSession := config.streamsPerSession * broadcastsPerStream
+
+			hub := NewHub(config.hubPoolSize)
+			node := NewMockNode()
+
+			go hub.Run()
+			defer hub.Shutdown()
+
+			var wg sync.WaitGroup
+			var streams []string
+
+			for i := 0; i < config.totalStreams; i++ {
+				stream := fmt.Sprintf("stream_%d", i)
+				streams = append(streams, stream)
+			}
+
+			for i := 0; i < config.totalSessions; i++ {
+				sid := fmt.Sprintf("%d", i)
+				session := NewMockSession(sid, &node)
+
+				rand.Seed(time.Now().Unix())
+
+				wg.Add(1)
+
+				go func() {
+					countdown := 0
+					for {
+						if countdown >= messagesPerSession {
+							wg.Done()
+							break
+						}
+
+						(session.ws.(MockConnection)).ReadIndifinitely()
+						countdown++
+					}
+				}()
+
+				hub.addSession(session)
+
+				for j := 0; j < config.streamsPerSession; j++ {
+					channel := fmt.Sprintf("test_channel%d", j)
+
+					stream := streams[rand.Intn(len(streams))]
+
+					hub.subscribeSession(sid, stream, channel)
+				}
+			}
+
+			b.ResetTimer()
+
+			for _, stream := range streams {
+				for i := 0; i < broadcastsPerStream; i++ {
+					hub.Broadcast(stream, config.payload)
+				}
+			}
+
+			wg.Wait()
+			b.StopTimer()
+		})
+	}
 }