+++ node/node.go
@@ -8,7 +8,9 @@ import (
 
 	"github.com/anycable/anycable-go/common"
 	"github.com/anycable/anycable-go/metrics"
+	"github.com/anycable/anycable-go/utils"
 	"github.com/apex/log"
+	"github.com/gorilla/websocket"
 )
 
 const (
@@ -35,6 +37,47 @@ type AppNode interface {
 	HandlePubSub(msg []byte)
 }
 
+// Connection represents underlying connection
+type Connection interface {
+	Write(msg []byte, deadline time.Time) error
+	Read() ([]byte, error)
+	Close(code int, reason string)
+}
+
+// WSConnection is a WebSocket implementation of Connection
+type WSConnection struct {
+	conn *websocket.Conn
+}
+
+// Write writes message to a WebSocket
+func (ws WSConnection) Write(msg []byte, deadline time.Time) error {
+	if err := ws.conn.SetWriteDeadline(deadline); err != nil {
+		return err
+	}
+
+	w, err := ws.conn.NextWriter(websocket.TextMessage)
+
+	if err != nil {
+		return err
+	}
+
+	if _, err = w.Write(msg); err != nil {
+		return err
+	}
+
+	return w.Close()
+}
+
+func (ws WSConnection) Read() ([]byte, error) {
+	_, message, err := ws.conn.ReadMessage()
+	return message, err
+}
+
+// Close sends close frame with a given code and a reason
+func (ws WSConnection) Close(code int, reason string) {
+	utils.CloseWS(ws.conn, code, reason)
+}
+
 // Node represents the whole application
 type Node struct {
 	Metrics *metrics.Metrics
@@ -57,7 +100,7 @@ func NewNode(controller Controller, metrics *metrics.Metrics, config *Config) *N
 		log:        log.WithFields(log.Fields{"context": "node"}),
 	}
 
-	node.hub = NewHub()
+	node.hub = NewHub(config.HubGopoolSize)
 
 	node.registerMetrics()
 
@@ -65,9 +108,11 @@ func NewNode(controller Controller, metrics *metrics.Metrics, config *Config) *N
 }
 
 // Start runs all the required goroutines
-func (n *Node) Start() {
+func (n *Node) Start() error {
 	go n.hub.Run()
 	go n.collectStats()
+
+	return nil
 }
 
 // SetDisconnector set disconnector for the node
@@ -135,6 +180,7 @@ func (n *Node) Shutdown() {
 			for _, session := range n.hub.sessions {
 				session.Send(disconnectMessage)
 				session.Disconnect("Shutdown", CloseGoingAway)
+				session.Flush()
 			}
 
 			n.log.Info("All active connections closed")
@@ -178,6 +224,12 @@ func (n *Node) Authenticate(s *Session) (err error) {
 	if res != nil {
 		n.handleCallReply(s, res.ToCallResult())
 	}
+
+	if err != nil {
+		s.Disconnect("Auth Error", CloseInternalServerErr)
+	}
+
+	s.Flush()
 
 	return
 }
@@ -316,6 +368,8 @@ func transmit(s *Session, transmissions []string) {
 }
 
 func (n *Node) handleCommandReply(s *Session, msg *common.Message, reply *common.CommandResult) {
+	defer s.Flush()
+
 	if reply.Disconnect {
 		defer s.Disconnect("Command Failed", CloseAbnormalClosure)
 	}